# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blossomer GTM API v2",
...
    "contact": {
      "name": "Blossomer GTM API Support",
      "email": "support@blossomer.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.blossomer.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/personas": {
      "post": {
        "tags": [
          "Personas",
          "Personas",
          "AI"
        ],
        "summary": "Generate Target Persona Profile (attributes, buying signals, rationale)",
        "description": "Generate a target persona profile for authenticated users (Stack Auth JWT required).",
        "operationId": "prod_generate_target_persona_api_personas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetPersonaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A structured target persona profile for the given company context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetPersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/personas-crud": {
      "post": {
        "tags": [
          "Personas"
        ],
        "summary": "Create Persona",
        "description": "Create a new persona for an account.\n\n- **account_id**: Account ID (must be owned by authenticated user via company)\n- **name**: Persona name (required, max 255 chars)\n- **persona_data**: JSON data with demographics, use cases, goals, objections",
        "operationId": "create_persona_api_personas_crud_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Account ID to create persona for",
              "title": "Account Id"
            },
            "description": "Account ID to create persona for"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Personas"
        ],
        "summary": "Get Personas",
        "description": "Get all personas for an account.\n\n- **account_id**: Account ID (must be owned by authenticated user via company)\n- **skip**: Number of personas to skip (for pagination)\n- **limit**: Maximum number of personas to return (1-1000)",
        "operationId": "get_personas_api_personas_crud_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Account ID to get personas for",
              "title": "Account Id"
            },
            "description": "Account ID to get personas for"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of personas to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of personas to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of personas to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of personas to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonaResponse"
                  },
                  "title": "Response Get Personas Api Personas Crud Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/personas-crud/{persona_id}": {
      "get": {
        "tags": [
          "Personas"
        ],
        "summary": "Get Persona",
        "description": "Get a specific persona by ID.\n\nOnly returns personas owned by the authenticated user (via account->company).",
        "operationId": "get_persona_api_personas_crud__persona_id__get",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Persona Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Personas"
        ],
        "summary": "Update Persona",
        "description": "Update a persona.\n\nOnly updates personas owned by the authenticated user (via account->company).\nAll fields are optional - only provided fields will be updated.",
        "operationId": "update_persona_api_personas_crud__persona_id__put",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Persona Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Personas"
        ],
        "summary": "Delete Persona",
        "description": "Delete a persona and all related campaigns.\n\nOnly deletes personas owned by the authenticated user (via account->company).\nThis will cascade delete all campaigns.",
        "operationId": "delete_persona_api_personas_crud__persona_id__delete",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Persona Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/personas-crud/{persona_id}/relations": {
      "get": {
        "tags": [
          "Personas"
        ],
        "summary": "Get Persona With Relations",
        "description": "Get a persona with all related campaigns.\n\nOnly returns personas owned by the authenticated user (via account->company).\nIncludes nested campaign data.",
        "operationId": "get_persona_with_relations_api_personas_crud__persona_id__relations_get",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Persona Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaWithRelations"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuyingSignal": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Concise signal name (3-5 words)"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "1-2 sentences explaining why this signal indicates buying readiness"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Company Data|Website|Tech Stack|News|Social Media|Other"
          },
          "priority": {
            "$ref": "#/components/schemas/PriorityEnum",
            "description": "Low|Medium|High"
          },
          "detection_method": {
            "type": "string",
            "title": "Detection Method",
            "description": "Specific Clay enrichment or data source"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "type",
          "priority",
          "detection_method"
        ],
        "title": "BuyingSignal"
      },
...
      "CampaignResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "campaign_type": {
            "type": "string",
            "maxLength": 50,
            "title": "Campaign Type",
            "description": "Campaign type: email, linkedin, cold_call, ad"
          },
          "campaign_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Campaign Data",
            "description": "Campaign data including subject_line, content, segments, alternatives, configuration"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Account Id"
          },
          "persona_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Persona Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "campaign_type",
          "campaign_data",
          "id",
          "account_id",
          "persona_id",
          "created_at",
          "updated_at"
        ],
        "title": "CampaignResponse",
        "description": "Schema for campaign responses."
      },
...
      "Demographics": {
        "properties": {
          "job_titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Job Titles",
            "description": "2-4 likely job titles this person would hold"
          },
          "departments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Departments",
            "description": "The department(s) they likely belong to"
          },
          "seniority": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seniority",
            "description": "Seniority levels: Entry|C-Suite|Senior Manager|Manager|VP|Founder/CEO"
          },
          "buying_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Buying Roles",
            "description": "Buying roles: Technical Buyers|Economic Buyers|Decision Maker|Champion|End-User|Blocker|Executive Sponsor|Legal and Compliance|Budget Holder"
          },
          "job_description_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Job Description Keywords",
            "description": "3-5 key words expected in job description describing day-to-day activities"
          }
        },
        "type": "object",
        "required": [
          "job_titles",
          "departments",
          "seniority",
          "buying_roles",
          "job_description_keywords"
        ],
        "title": "Demographics",
        "description": "Demographics model for target persona."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PersonaCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "persona_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Persona Data",
            "description": "Persona data including demographics, use cases, buying signals, objections, goals"
          }
        },
        "type": "object",
        "required": [
          "name",
          "persona_data"
        ],
        "title": "PersonaCreate",
        "description": "Schema for creating a new persona.",
        "example": {
          "name": "VP of Engineering",
          "persona_data": {
            "demographics": {
              "buying_roles": [
                "Decision Maker",
                "Technical Buyer",
                "Economic Buyer"
              ],
              "departments": [
                "Engineering",
                "Technology"
              ],
              "job_description_keywords": [
                "team leadership",
                "technical strategy",
                "developer productivity",
                "scaling",
                "automation"
              ],
              "job_titles": [
                "VP Engineering",
                "Head of Engineering",
                "Engineering Director"
              ],
              "seniority": [
                "VP",
                "Director",
                "Senior Manager"
              ]
            },
            "goals": [
              "Improve team productivity and delivery speed",
              "Reduce technical debt and improve code quality",
              "Scale engineering processes as team grows",
              "Minimize time spent on manual, repetitive tasks"
            ],
            "objections": [
              "Concerned about integration complexity with existing tools",
              "Budget approval process may be lengthy",
              "Team resistance to changing established workflows"
            ],
            "use_cases": [
              {
                "capability": "AI-powered code review that catches issues early and provides instant feedback",
                "desired_outcome": "Faster development cycles with maintained code quality",
                "pain_points": "Manual code reviews slow down development cycles and create bottlenecks",
                "use_case": "Code review automation"
              },
              {
                "capability": "Intelligent pipeline optimization that reduces build times by 40%",
                "desired_outcome": "Reliable, fast deployments that don't block development",
                "pain_points": "Build pipelines are slow and unreliable, causing deployment delays",
                "use_case": "CI/CD optimization"
              }
            ]
          }
        }
      },
      "PersonaResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "persona_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Persona Data",
            "description": "Persona data including demographics, use cases, buying signals, objections, goals"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Account Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "persona_data",
          "id",
          "account_id",
          "created_at",
          "updated_at"
        ],
        "title": "PersonaResponse",
        "description": "Schema for persona responses."
      },
      "PersonaUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "persona_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Data"
          }
        },
        "type": "object",
        "title": "PersonaUpdate",
        "description": "Schema for updating persona information."
      },
      "PersonaWithRelations": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "persona_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Persona Data",
            "description": "Persona data including demographics, use cases, buying signals, objections, goals"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Account Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "campaigns": {
            "items": {
              "$ref": "#/components/schemas/CampaignResponse"
            },
            "type": "array",
            "title": "Campaigns",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "persona_data",
          "id",
          "account_id",
          "created_at",
          "updated_at"
        ],
        "title": "PersonaWithRelations",
        "description": "Persona schema with related campaigns."
      },
...
      "PriorityEnum": {
        "type": "string",
        "enum": [
          "Low",
          "Medium",
          "High"
        ],
        "title": "PriorityEnum"
      },
...
      "TargetPersonaRequest": {
        "properties": {
          "website_url": {
            "type": "string",
            "title": "Website Url",
            "description": "Company website or landing page URL"
          },
          "persona_profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Profile Name",
            "description": "Name of the target persona profile (e.g., 'VP of Engineering', 'Head of Customer Success', 'IT Director')"
          },
          "hypothesis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypothesis",
            "description": "User's hypothesis about why this persona is ideal for the solution"
          },
          "additional_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Context",
            "description": "Additional context or specifications about the target persona"
          },
          "company_context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Context",
            "description": "Structured context about the analyzed company/product"
          },
          "target_account_context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Account Context",
            "description": "Target account profile context - the ideal customer company type this persona works for"
          }
        },
        "type": "object",
        "required": [
          "website_url"
        ],
        "title": "TargetPersonaRequest"
      },
      "TargetPersonaResponse": {
        "properties": {
          "target_persona_name": {
            "type": "string",
            "title": "Target Persona Name",
            "description": "Short descriptive name for this persona segment"
          },
          "target_persona_description": {
            "type": "string",
            "title": "Target Persona Description",
            "description": "1-2 sentences: who they are and why they need this solution"
          },
          "target_persona_rationale": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Persona Rationale",
            "description": "3-5 bullets explaining the overall logic behind targeting this persona"
          },
          "demographics": {
            "$ref": "#/components/schemas/Demographics",
            "description": "Demographics and targeting attributes"
          },
          "use_cases": {
            "items": {
              "$ref": "#/components/schemas/UseCase"
            },
            "type": "array",
            "title": "Use Cases",
            "description": "3-4 use cases following logical progression"
          },
          "buying_signals": {
            "items": {
              "$ref": "#/components/schemas/BuyingSignal"
            },
            "type": "array",
            "title": "Buying Signals",
            "description": "Observable buying signals with detection methods"
          },
          "buying_signals_rationale": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Buying Signals Rationale",
            "description": "3-5 bullets explaining buying signal logic"
          },
          "objections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Objections",
            "description": "3 bullets summarizing common concerns about adopting this solution"
          },
          "goals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Goals",
            "description": "3-5 bullets explaining business objectives this product can help with"
          },
          "purchase_journey": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Purchase Journey",
            "description": "3-6 bullet points highlighting path from awareness to purchase"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Analysis metadata and quality scores"
          }
        },
        "type": "object",
        "required": [
          "target_persona_name",
          "target_persona_description",
          "target_persona_rationale",
          "demographics",
          "use_cases",
          "buying_signals",
          "buying_signals_rationale",
          "objections",
          "goals",
          "purchase_journey",
          "metadata"
        ],
        "title": "TargetPersonaResponse",
        "description": "Response model for the /customers/target_personas endpoint (matches new prompt output)."
      },
...
      "UseCase": {
        "properties": {
          "use_case": {
            "type": "string",
            "title": "Use Case",
            "description": "3-5 word description of the use case or workflow this product impacts"
          },
          "pain_points": {
            "type": "string",
            "title": "Pain Points",
            "description": "1 sentence description of the pain or inefficiency associated with this pain point"
          },
          "capability": {
            "type": "string",
            "title": "Capability",
            "description": "1 sentence description of the capability the product has that can fix this pain point"
          },
          "desired_outcome": {
            "type": "string",
            "title": "Desired Outcome",
            "description": "The desired outcome the persona hopes to achieve using this product"
          }
        },
        "type": "object",
        "required": [
          "use_case",
          "pain_points",
          "capability",
          "desired_outcome"
        ],
        "title": "UseCase",
        "description": "Individual use case model for target persona."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}